moved translatedword to own svelte component
