+++ src/lib/translated-text-window-props.types.ts
@@ -1,0 +1,6 @@
+import type {TranslatedText} from "./utils/translate-text";
+
+export interface TranslatedTextWindowProps {
+    translatedText: TranslatedText[];
+    translateEnglishToDovahzul: boolean;
+}
